+++ src/entities/club/api/clubQueries.ts
@@ -5,17 +5,10 @@ import {
   useMutation,
   useQueryClient,
 } from "@tanstack/react-query";
-import { instance } from "@/shared/api/instance";
 import { getClubs } from "./getClubs";
 import { getClubDetail } from "./getClub";
-import { getClubForm } from "./getClubForm";
-import { getClubApplications } from "./getClubApplications";
 import { getClubOpeningRequests } from "./getClubOpeningRequests";
 import { getClubOpeningStatus } from "./getClubOpeningStatus";
-import type {
-  ClubApplicationRequest,
-  CreateClubFormRequest,
-} from "../model/club";
 import { patchClubApproval } from "./patchClubApproval";
 import { postClub } from "./postClub";
 import { deleteClub } from "./deleteClub";
@@ -35,18 +28,6 @@ export const clubQueries = {
       queryFn: () => getClubDetail(id),
     }),
 
-  form: (clubId: number) =>
-    queryOptions({
-      queryKey: ["club", "form", clubId],
-      queryFn: () => getClubForm(clubId),
-    }),
-
-  applicationList: (clubId: number) =>
-    queryOptions({
-      queryKey: ["club", "applications", clubId],
-      queryFn: () => getClubApplications(clubId),
-    }),
-
   openingRequests: () =>
     queryOptions({
       queryKey: ["club", "opening-requests"],
@@ -58,32 +39,6 @@ export const clubQueries = {
       queryKey: ["club", "opening-status"],
       queryFn: getClubOpeningStatus,
     }),
-} as const;
-
-export const clubMutations = {
-  createForm: (clubId: number, body: CreateClubFormRequest) =>
-    instance.post(`/clubs/${clubId}/forms`, body),
-
-  updateForm: (clubId: number, body: CreateClubFormRequest) =>
-    instance.put(`/clubs/${clubId}/forms`, body),
-
-  applyClub: (clubId: number, body: ClubApplicationRequest) =>
-    instance.post(`/clubs/${clubId}/applications`, body),
-
-  applyAutonomousClub: (clubId: number) =>
-    instance.post(`/clubs/${clubId}/autonomous/applications`),
-
-  approveApplication: (clubId: number, userId: number) =>
-    instance.patch(`/clubs/${clubId}/applications/${userId}`),
-
-  transferLeader: (clubId: number, targetUserId: number) =>
-    instance.patch(`/clubs/${clubId}/transfer/${targetUserId}`),
-
-  inviteMember: (clubId: number, userId: number) =>
-    instance.post(`/clubs/${clubId}/member/${userId}`),
-
-  exileMember: (clubId: number, userId: number) =>
-    instance.delete(`/clubs/${clubId}/member/exile/${userId}`),
 } as const;
 
 export const usePatchClubApproval = () => {
