+++ src/entities/club/model/club.ts
@@ -67,96 +67,6 @@ export interface ClubDetailResponse {
   isLeader: boolean;
 }
 
-export type ClubFormFieldType =
-  | "TEXT"
-  | "TEXTAREA"
-  | "RADIO"
-  | "CHECKBOX"
-  | "DROPDOWN";
-
-export interface ClubFormFieldOption {
-  optionId: number;
-  label: string;
-  value: string;
-}
-
-export interface ClubFormField {
-  fieldId: number;
-  label: string;
-  description?: string;
-  fieldType: ClubFormFieldType;
-  order: number;
-  required: boolean;
-  options: ClubFormFieldOption[];
-}
-
-export interface ClubForm {
-  formId: number;
-  title: string;
-  description?: string;
-  fields: ClubFormField[];
-}
-
-export interface CreateClubFormFieldOptionRequest {
-  label: string;
-  value: string;
-}
-
-export interface CreateClubFormFieldRequest {
-  label: string;
-  description?: string;
-  fieldType: ClubFormFieldType;
-  order: number;
-  required: boolean;
-  options?: CreateClubFormFieldOptionRequest[];
-}
-
-export interface CreateClubFormRequest {
-  title: string;
-  description?: string;
-  fields: CreateClubFormFieldRequest[];
-}
-
-export interface CreateClubFormResponse {
-  formId: number;
-}
-
-export interface ClubApplicationAnswerRequest {
-  fieldId: number;
-  value?: string;
-}
-
-export interface ClubApplicationRequest {
-  answers: ClubApplicationAnswerRequest[];
-}
-
-export interface ClubApplicationResponse {
-  applicationId: number;
-}
-
-export interface ClubApplicantInfo {
-  id: number;
-  name: string;
-  studentNumber: number;
-}
-
-export interface ClubApplicationAnswer {
-  fieldId: number;
-  label: string;
-  value?: string;
-}
-
-export interface ClubApplicationSummary {
-  submissionId: number;
-  applicant: ClubApplicantInfo;
-  submittedAt: string;
-  answers: ClubApplicationAnswer[];
-}
-
-export interface ClubApplicationListResponse {
-  applications: ClubApplicationSummary[];
-}
-
 export interface ClubOpeningStatus {
   isOpened: boolean;
   startDate: string | null;
